rename ComponentModel occurrences to ModelComponent to be consistent
