+++ .eslintignore-circle
@@ -110,7 +110,7 @@ src/scope/network/ssh/ssh.js
 src/hooks/hooks-manager.js
 src/scope/component-remove.js
 src/extensions/extension.js
-src/scope/models/component.js
+src/scope/models/model-component.js
 src/consumer/consumer.js
 e2e/e2e-helper.js
 src/consumer/component/consumer-component.js
